refactor: switch to using ServerManager and CheckingServerManager
