+++ src/main/java/com/atlauncher/gui/tabs/ServersForCheckerTab.java
@@ -34,9 +34,9 @@ import javax.swing.JPopupMenu;
 import javax.swing.ListSelectionModel;
 import javax.swing.event.ListSelectionListener;
 
-import com.atlauncher.App;
 import com.atlauncher.data.MinecraftServer;
 import com.atlauncher.gui.dialogs.AddEditServerForCheckerDialog;
+import com.atlauncher.managers.CheckingServersManager;
 
 import org.mini2Dx.gettext.GetText;
 
@@ -59,7 +59,7 @@ public class ServersForCheckerTab extends JPanel implements ActionListener {
         CONTEXT_MENU.add(DELETE_BUTTON);
 
         listModel = new DefaultListModel<>();
-        for (MinecraftServer server : App.launcher.getCheckingServers()) {
+        for (MinecraftServer server : CheckingServersManager.getCheckingServers()) {
             listModel.addElement(server);
         }
         serverList = new JList(listModel);
@@ -121,7 +121,7 @@ public class ServersForCheckerTab extends JPanel implements ActionListener {
     public void deleteSelectedElement() {
         if (serverList.getSelectedIndex() != -1) {
             MinecraftServer selectedValue = ((MinecraftServer) serverList.getSelectedValue());
-            App.launcher.removeCheckingServer(selectedValue);
+            CheckingServersManager.removeCheckingServer(selectedValue);
             listModel.removeElement(selectedValue);
             reloadServers();
         }
@@ -129,7 +129,7 @@ public class ServersForCheckerTab extends JPanel implements ActionListener {
 
     public void reloadServers() {
         listModel.removeAllElements();
-        for (MinecraftServer server : App.launcher.getCheckingServers()) {
+        for (MinecraftServer server : CheckingServersManager.getCheckingServers()) {
             listModel.addElement(server);
         }
     }
